delete warnings when compiling

+++ awadb/db_engine/c_api/api_data/gamma_response.cc
@@ -22,7 +22,7 @@ void ResultItem::GetVecData(const std::string &fid_name, std::vector<float> &res
     if (fid_name == iter_name)  {
       int v = 0;
       memcpy((void *)&v, this->values[i].c_str(), sizeof(int)); 
-      for (int j = 0; j < v / sizeof(float); j++)  {
+      for (size_t j = 0; j < v / sizeof(float); j++)  {
         float f_value;
 	memcpy((void *)&f_value, (void *)(this->values[i].c_str() + sizeof(int) + j * sizeof(float)), sizeof(float));
 	result_vec.push_back(f_value);
